requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.43.0
grpcio==1.76.0
grpcio-tools==1.76.0
idna==3.11
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
